Make server success code match webpush (#25)
Changing 202 to 201 is a very insignificant change since 2xx is required to be accepted (and most webpush implementations accept 2xx anyway https://github.com/UnifiedPush/wishlist/issues/15#issuecomment-1086995693)

However, it is a minor way in which UnifiedPush can be similar to WebPush and could matter to pickier implementations. 201 is required in webpush. https://datatracker.ietf.org/doc/html/rfc8030#section-5